Moved analysis and statistics integration tests to different folder
All that was added was:
     .WithAuthentication()
     .WithAuthenticationStores()
